server/requirements.txt
openenv-core
pydantic>=2.0
fastapi
uvicorn[standard]
websockets
httpx
pyyaml
python-dotenv

# Training stack (heavy; pinned cautiously)
transformers==4.46.0
torch
trl==0.13.0
accelerate
datasets

# Dev
pytest
